Move the `app-common` module under the `@bayou` namespace.
