+++ client/package.json
@@ -17,7 +17,7 @@
 
   "dependencies": {
     "@bayou/api-common": "local",
-    "app-common": "local",
+    "@bayou/app-common": "local",
     "doc-client": "local",
     "env-client": "local",
     "@bayou/see-all-client": "local",
