+++ local-modules/doc-server/package.json
@@ -5,7 +5,7 @@
   "dependencies": {
     "weak": "^1.0.1",
 
-    "app-common": "local",
+    "@bayou/app-common": "local",
     "@bayou/codec": "local",
     "@bayou/doc-common": "local",
     "env-server": "local",
