+++ local-modules/doc-server/tests/test_BaseComplexMember.js
@@ -7,7 +7,7 @@ import { describe, it } from 'mocha';
 
 import { BaseComplexMember } from 'doc-server';
 
-import { TheModule as appCommon_TheModule } from 'app-common';
+import { TheModule as appCommon_TheModule } from '@bayou/app-common';
 import { FileAccess } from 'doc-server';
 import { MockFile } from 'file-store/mocks';
 import { MockLogger } from '@bayou/see-all/mocks';
